requirements.txt
pandas==2.1.4
numpy==1.26.3
scikit-learn==1.3.2
pytest==7.4.4
apache-airflow-providers-docker==3.9.0